+++ config/connection.js
@@ -1,7 +1,7 @@
 const { connect, connection } = require('mongoose');
 
 // Wrap Mongoose around local connection to MongoDB
-mongoose.connect('mongodb://127.0.0.1:27017/mygroceryDB');
+connect('mongodb://127.0.0.1:27017/mygroceryDB');
 
 // Export connection 
-module.exports = mongoose.connection;
+module.exports = connection;
