--- index.js
@@ -1,18 +1,0 @@
-const express = require('express');
-const db = require('./config/connection');
-const routes = require('./routes');
-
-const cwd = process.cwd();
-// require model
-const { Item } = require('./models');
-const PORT = process.env.PORT || 3001;
-const app = express();
-
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-
-db.once('open', () => {
-  app.listen(PORT, () => {
-    console.log(`API server running on port ${PORT}!`);
-  });
-});
